+++ R/inference.R
@@ -29,7 +29,7 @@ greta_stash$numerical_messages <- c(
 #'   rest are discarded
 #' @param warmup number of samples to spend warming up the mcmc sampler (moving
 #'   chains toward the highest density area and tuning sampler hyperparameters).
-#' @param chains number of MCMC chains to run.
+#' @param chains number of MCMC chains to run. Default is 2. We recommend using more chains as this helps improve convergence. However the number of chains specified can increase the CPU load, so we have to set a lower default value.
 #' @param n_cores the maximum number of CPU cores used by each sampler (see
 #'   details). If NULL (default), it sets them to 2 cores.
 #' @param verbose whether to print progress information to the console
@@ -212,7 +212,7 @@ mcmc <- function(
   n_samples = 1000,
   thin = 1,
   warmup = 1000,
-  chains = 4,
+  chains = 2,
   n_cores = NULL,
   verbose = TRUE,
   pb_update = 50,