+++ NEWS.md
@@ -6,6 +6,7 @@
 - Add warmup information to MCMC print method (#652, resolved by #755).
 - Add more options to level of detail in `greta_sitrep()` with "verbosity" argument. There are three levels, "minimal" (default), "detailed", and "quiet". (#612, resolved by #679).
 - Use `.batch_size` instead of `batch_size` internally, to avoid rare name clash errors (#634).
+* Resolve issues with Tensorflow version in DESCRIPTION (no longer can specify == 2.16.0, must be >= 2.16.0).
 
 # greta 0.5.0